backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.1
requests==2.31.0
django-cors-headers==4.3.1
